pyproject.toml
[tool.poetry]
name = "cooling-centers"
version = "1.4.0"
description = ""
authors = [
    "pythoninthegrass <4097471+pythoninthegrass@users.noreply.github.com>",
    "Alex Ayon <jayonloza@gmail.com>"
]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
beautifulsoup4 = "^4.12.3"
fastapi = "^0.111.0"
folium = "^0.14.0"
geopy = {extras = ["aiohttp"], version = "^2.4.1"}
hishel = {extras = ["sqlite"], version = "^0.0.29"}
httpx = "^0.27.0"
lxml = "^5.2.2"
numpy = "^1.26.4"
pandas = "^2.2.2"
pdftotree = "^0.5.0"
python-decouple = "^3.8"
rich = "^13.7.1"
strawberry-graphql = {extras = ["debug-server", "fastapi"], version = "^0.235.0"}
tabula-py = {extras = ["jpype"], version = "^2.9.3"}
uvicorn = "^0.30.1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.5.3"
hypothesis = {extras = ["cli"], version = "^6.103.2"}
icecream = "^2.1.3"
ipython = "^8.25.0"
playwright = "^1.45.0"
poetry-plugin-export = "^1.8.0"
poetry-plugin-up = "^0.7.1"
pyclean = "^3.0.0"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.3"
pytest-cov = "^5.0.0"
pytest-datafiles = "^3.0.0"
pytest-xdist = "^3.6.1"
ruff = "^0.4.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [
    ".cache",
    ".eggs",
    ".git",
    ".github",
    ".gitmodules",
    ".pytest",
    ".pytest_cache",
    ".venv",
    ".vscode",
    "__pycache__",
    "build",
    "dist",
    "node_modules"
]

[tool.uv.pip]
index-url = "https://pypi.org/simple"

[tool.ruff]
# Fix without reporting on leftover violations
fix-only = true

# Enumerate all fixed violations
show-fixes = true

# Indent width (default: 4)
indent-width = 4

# Black (default: 88)
line-length = 130

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    "dist",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    "__pycache__",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Assume Python 3.11.
target-version = "py311"

[tool.ruff.format]
# Use spaces instead of tabs
indent-style = "space"

# Use `\n` line endings for all files
line-ending = "lf"

# Set quote style for strings
quote-style = "preserve"

[tool.ruff.lint]
# Enable pycodestyle (`E`), Pyflakes (`F`), and isort (`I`) by default.
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001",
    "I002"
]
# Skip unused variable rules
ignore = []
extend-ignore = ["D203", "E203", "E251", "E266", "E401", "E402", "E501", "F401", "F403"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]

# unfixable = []

[tool.ruff.lint.isort]
from-first = false
lines-after-imports = 1
lines-between-types = 0
no-sections = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
